Step 6: Migrate analysis and validation to modular structure
- Created src/dtu_analyzer/analysis/analyzer.py
  - Moved from root analyzer.py
  - Refactored into separate functions (process_courses, generate_extension_data, etc.)
  - Updated imports to use config, logger, and prepender modules
  - Uses config.paths for file locations
  - Added type hints throughout

- Created src/dtu_analyzer/validation/validator.py
  - Moved from root validator.py
  - Updated imports to use config and logger modules
  - Uses config.paths for default file location
  - Maintained all validation logic

- Created backward-compatible wrappers:
  - analyzer.py → imports from new location
  - validator.py → imports from new location

- Updated module __init__.py files:
  - analysis/__init__.py exports main functions
  - validation/__init__.py exports validator classes

All existing scripts and workflows continue to work via wrappers
